Actualizar

+++ appsireca/perfil.py
@@ -172,9 +172,13 @@ def view(request):
                                 accesomodulo.save()
 
                     # verificar si se quito un modulo del perfil
-                    ideliminados = list(set(listamoduloantes) - set(list(listamoduloingreso)))
-                    ModuloPerfil.objects.filter(perfil=perfil, modulo__id__in=ideliminados).delete()
-                    AccesoModulo.objects.filter(perfilpersona__perfil=perfil, modulo__id__in=ideliminados).delete()
+                    ids_antes = [m.id for m in listamoduloantes]
+                    ids_ingreso = [m for m in listamoduloingreso]
+                    ideliminados = list(set(ids_antes) - set(ids_ingreso))
+
+                    if ideliminados:
+                        ModuloPerfil.objects.filter(perfil=perfil, modulo__id__in=ideliminados).delete()
+                        AccesoModulo.objects.filter(perfilpersona__perfil=perfil, modulo__id__in=ideliminados).delete()
 
 
                     data['result'] = 'ok'
